Extract SpotifyService from SetlistService
