+++ SpotSet.Api.Tests/Services/SetlistFmServiceTest.cs
@@ -64,66 +64,5 @@ namespace SpotSet.Api.Tests.Services
 
             Assert.Null(result);
         }
-
-        [Fact]
-        public async void SpotifyRequestReturnsASpotifyModelWhenCalledWithSetlistModel()
-        {
-            var song1 = new Song
-            {
-                Name = "Song Title"
-            };
-            
-            var song2 = new Song
-            {
-                Name = "Another Song Title"
-            };
-            
-            var setWithTwoSongs = new Set
-            {
-                Song = new List<Song>
-                {
-                    song1, song2
-                }
-            };
-
-            var newSets = new Sets
-            {
-                Set = new List<Set>
-                {
-                    setWithTwoSongs
-                }
-            };
-
-            var newSetlist = new SetlistDto
-            {
-                Id = "setlistId",
-                EventDate = "01-07-2019",
-                Artist = new Artist { Name = "Artist" },
-                Venue = new Venue { Name = "Venue" },
-                Sets = newSets
-            };
-            
-            var successSetlistService = TestSetup.CreateSetlistServiceWithMocks(HttpStatusCode.OK, newSetlist);
-            var result = await successSetlistService.SpotifyRequest(newSetlist);
-
-            Assert.IsType<SpotifyTracksModel>(result);
-        }
-
-        [Fact]
-        public async void SpotifyRequestReturnsEmptySpotifyTracksModelWhenCalledWithAnInvalidSetlistModel()
-        {
-            var newSetlist = new SetlistDto
-            {
-                Id = "setlistId",
-                EventDate = "01-07-2019",
-                Artist = null,
-                Venue = null,
-                Sets = new Sets { Set = new List<Set>() }
-            };
-            var successSetlistService = TestSetup.CreateSetlistServiceWithMocks(HttpStatusCode.OK, newSetlist);
-            var result = await successSetlistService.SpotifyRequest(newSetlist);
-            
-            Assert.Equal(0, result.SpotifyTracks.Count);
-        }
     }
 }
