+++ SpotSet.Api/Services/ISetlistService.cs
@@ -6,6 +6,5 @@ namespace SpotSet.Api.Services
     public interface ISetlistService
     {
         Task<SpotSetDto> GetSetlist(string setlistId);
-        Task<SpotifyTracksModel> SpotifyRequest(SetlistDto setlistmodel);
     }
 }
