+++ SpotSet.Api/Services/ISpotifyService.cs
@@ -1,0 +1,10 @@
+using System.Threading.Tasks;
+using SpotSet.Api.Models;
+
+namespace SpotSet.Api.Services
+{
+    public interface ISpotifyService
+    {
+        Task<SpotifyTracksModel> SpotifyRequest(SetlistDto setlistmodel);
+    }
+}
